requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic-settings
gunicorn